Migrate the Cfg pane to typescript
